+++ app/modusoperandi.cpp
@@ -21,8 +21,9 @@ ModusOperandi::ModusOperandi(Config::Local &local_cfg, SlidingBanner *banner, QO
   connect(&mpd_client, &MpdClient::Client::error, this, &ModusOperandi::mpdLost);
   connect(&mpd_client, &MpdClient::Client::disconnected, this, &ModusOperandi::mpdLost);
 
-  connect(&mpd_client, &MpdClient::Client::error, [=] {
-    banner->showMessage(tr("mpd connection error"), SlidingBanner::BannerType::Error);
+  connect(&mpd_client, &MpdClient::Client::error, [=](const QUrl &url, const QString &message) {
+    Q_UNUSED(url);
+    banner->showMessage(tr("mpd connection error") + "\n" + message, SlidingBanner::BannerType::Error);
   });
   connect(&mpd_client, &MpdClient::Client::connected, [=] {
     banner->showMessage(tr("mpd connected"), SlidingBanner::BannerType::Success, 3456);