+++ app/mpd_client/connection.h
@@ -63,7 +63,7 @@ namespace MpdClient {
   signals:
     void connected(const QUrl &url);
     void disconnected(const QUrl &url);
-    void error(const QUrl &url);
+    void error(const QUrl &url, const QString &message = "");
     void idleEvent(mpd_idle event);
 
   private slots: